Integrate defaults with TestExclude
