--- lib/defaults.js
@@ -1,24 +1,0 @@
-const { isBare } = require('which-runtime')
-const process = require(isBare ? 'bare-process' : 'process')
-
-module.exports = {
-  testExclude: {
-    cwd: process.cwd(),
-    exclude: [
-      'coverage/**',
-      'packages/*/test{,s}/**',
-      '**/*.d.ts',
-      'test{,s}/**',
-      'test{,-*}.{js,cjs,mjs,ts,tsx,jsx}',
-      '**/*{.,-}test.{js,cjs,mjs,ts,tsx,jsx}',
-      '**/__tests__/**',
-      '**/{ava,babel,nyc}.config.{js,cjs,mjs}',
-      '**/jest.config.{js,cjs,mjs,ts}',
-      '**/{karma,rollup,webpack}.config.js',
-      '**/.{eslint,mocha}rc.{js,cjs}'
-    ],
-    excludeNodeModules: true,
-    include: [],
-    extension: ['.js', '.cjs', '.mjs', '.ts', '.tsx', '.jsx']
-  }
-}
