+++ lib/test-exclude.js
@@ -1,7 +1,23 @@
 const { isBare, isWindows } = require('which-runtime')
 const path = require(isBare ? 'bare-path' : 'path')
+const process = require(isBare ? 'bare-process' : 'process')
 const picomatch = require('picomatch')
-const defaults = require('./defaults')
+
+const DEFAULT_EXCLUDES = [
+  'coverage/**',
+  'packages/*/test{,s}/**',
+  '**/*.d.ts',
+  'test{,s}/**',
+  'test{,-*}.{js,cjs,mjs,ts,tsx,jsx}',
+  '**/*{.,-}test.{js,cjs,mjs,ts,tsx,jsx}',
+  '**/__tests__/**',
+  '**/{ava,babel,nyc}.config.{js,cjs,mjs}',
+  '**/jest.config.{js,cjs,mjs,ts}',
+  '**/{karma,rollup,webpack}.config.js',
+  '**/.{eslint,mocha}rc.{js,cjs}'
+]
+
+const DEFAULT_EXTENSIONS = ['.js', '.cjs', '.mjs', '.ts', '.tsx', '.jsx']
 
 function isOutsideDir (dir, filename) {
   return isWindows
@@ -12,9 +28,12 @@ function isOutsideDir (dir, filename) {
 // Based off https://github.com/istanbuljs/test-exclude/blob/master/index.js with glob-related code removed and minimatch replaced with picomatch
 class TestExclude {
   constructor (opts = {}) {
-    Object.assign(this, { relativePath: true }, defaults.testExclude)
-
-    for (const [name, value] of Object.entries(opts)) if (value !== undefined) this[name] = value
+    this.cwd = opts.cwd || process.cwd()
+    this.include = opts.include || []
+    this.exclude = opts.exclude || DEFAULT_EXCLUDES
+    this.extension = opts.extension || DEFAULT_EXTENSIONS
+    this.excludeNodeModules = opts.excludeNodeModules || true
+    this.relativePath = opts.relativePath || true
 
     if (typeof this.include === 'string') this.include = [this.include]
     if (typeof this.exclude === 'string') this.exclude = [this.exclude]
