+++ lib/transformer.js
@@ -7,18 +7,11 @@ const v8ToIstanbul = require('v8-to-istanbul')
 const TestExclude = require('./test-exclude')
 const summarize = require('./summarize')
 const report = require('./report')
-const defaults = require('./defaults')
 
 class Transformer {
   constructor (opts = {}) {
     this.includeRelative = opts.includeRelative ?? false
-    this.exclude = new TestExclude({
-      exclude: defaults.testExclude.exclude,
-      include: [],
-      extension: defaults.testExclude.extension,
-      relativePath: true,
-      excludeNodeModules: true
-    })
+    this.exclude = new TestExclude()
     this.dir = opts.dir ?? 'coverage'
     this.reporters = opts.reporters ?? ['text', 'json']
     this.reporterOptions = opts.reporterOptions ?? {}
